corrections

+++ 0x00-hello_world/6-size.c
@@ -15,9 +15,9 @@ int main(void)
  long long int d;
  float f;
 
-	printf("Size of char:%lu",sizeof(c);
-	printf("Size of int:%lu",sizeof(a);
-	printf("Size of long long int:%lu",sizeof(d);
-	printf("Size of long int:%lu",sizeof(b);
-	printf("Size of float:%lu",sizeof(f);
+	printf("Size of char:%lu",(unsigned long)sizeof(c));
+	printf("Size of int:%lu",(unsigned long)sizeof(a));
+	printf("Size of long long int:%lu",(unsigned long)sizeof(d));
+	printf("Size of long int:%lu",(unsigned long)sizeof(b));
+	printf("Size of float:%lu",(unsigned long)sizeof(f));
 }
